Change struct to Float32 and add point selection

+++ IOS-inst-clstr/IOS-inst-clstr/Backend/communication.swift
@@ -16,13 +16,13 @@ public struct APiDataPack : Decodable{
     var throttlePercent : Int = 0;
     var dutyPercent : Int = 0;
     var pwmFrequency : Int = 0;
-    var rpm : Float64 = 0.0;
-    var torque : Float64 = 0.0;
-    var tempC : Float64 = 0.0;
-    var sourceVoltage : Float64 = 0.0;
-    var pwmCurrent : Float64 = 0.0;
-    var powerChange : Float64 = 0.0;
-    var voltageChange : Float64 = 0.0;
+    var rpm : Float32 = 0.0;
+    var torque : Float32 = 0.0;
+    var tempC : Float32 = 0.0;
+    var sourceVoltage : Float32 = 0.0;
+    var pwmCurrent : Float32 = 0.0;
+    var powerChange : Float32 = 0.0;
+    var voltageChange : Float32 = 0.0;
     // graphable data
     var mddStatus : Bool = false;
     var ocpStatus : Bool = false;
